test(es/websocket): fix test errors
